+++ src/@locale/components/index.js
@@ -1,12 +1,16 @@
 import buttons    from './buttons'
 import dropdowns  from './dropdowns'
+import modal      from './modal'
 import pagination from './pagination'
+import splitPanel from './splitPanel'
 
 const components =
 {
     buttons ,
     dropdowns,
+    modal ,
     pagination ,
+    splitPanel ,
 };
 
 export default components ;